#48 Move modules into inkly package and update imports
Reorganize code into the new inkly package: rename config.py -> inkly/config.py and ink_core.py -> inkly/ink_core.py, add inkly/__init__.py with __version__. Update imports in the CLI wrapper (ink), install.py, and tests to use inkly.config and inkly.ink_core. Remove obsolete test-lint.py. This namespaces project modules under inkly and keeps imports consistent.
