+++ inkly/__init__.py
@@ -1,0 +1,3 @@
+# inkly/__init__.py
+
+__version__ = "0.1.0"
