--- test-lint.py
@@ -1,6 +1,0 @@
-import random
-
-
-def test_random_number():
-    num = random.randint(1, 10)
-    print(f"Generated number: {num}")
